Recursive Digit Sum/Modulo by 9/number to base 10
This problem entails to find the final digit sum of a number n with an additional constraint that the number will be multiplied by a given integer k that can range from 1 to 100000. And the number can range up to infinity. The digit sum is the end result of repeatedly computing the sum of the digits until a single digit answer is obtained. The problem can be easily solved using the math rule: The digit sum of any number base k is equivalent to its remainder upon division by (k-1). And additionally using the intuition: the number of unique digits allowed in any base is One-Zero in that base and it is ten (10) in decimal system, the problem can be solved by finding the remainder when the input number n is divided by 9 using modulo function in python.