Max remainder from adding and dividing subarrays
We define the following:
A subarray of array a of length n is a contiguous segment from a[i] through a[j] where 0<=i<=j<n.
The sum of an array is the sum of its elements.
Given an n element array of integers, a, and an integer, m, determine the maximum value of the sum of any of its subarrays modulo m.
For example, Assume a=[1,2,3] and m=2. The following table lists all subarrays and their moduli:
	    sum	%2
[1]		1	1
[2]		2	0
[3]		3	1
[1,2]	3      1
[2,3]	5	1
[1,2,3]	6	0
The maximum modulus is 1

Approach to solve the challenge
We create a prefix array (similar to the idea of having a treeset in java) initally with 0 as values to the length of the given array a. Then find the remainder using the following math formulas and append it to the prefix.
(𝑎+𝑏)%m=(𝑎%m+𝑏%m)%m
(𝑎−𝑏)%m=(𝑎%m−𝑏%m)%m
After traversing the given array once and appending the prefix array with remainders it's time to find the subarrays. A significant point to note is that the subarrays can be expressed in the following way:
subarray[i,j]=(prefix[j]-prefix[i-1]+m)%m

Intuition behind:
Let's say we have found the modulus sum of array (according to the above formula) A0 to Ai, then a subarray within A0 to Ai will only produce a larger mod sum if the start of the subarray is greater than the start of the array A0 to Ai. Thus we don't have to calulate the mod sum of every possible subarray. Knowing this will help us to think along the lines of binary tree or binary search a sorted array. I have used bisect module to solve the challenge inspired by https://www.hackerrank.com/challenges/maximum-subarray-sum/forum