To find lonely integer in an array
You will be given an array of integers. All of the integers except one occur twice. That one is unique in the array.
Given an array of integers, find and print the unique element.
For example, a = [1,2,3,4,3,2,1], the unique element is 4.

Function Description
Complete the lonelyinteger function in the editor below. It should return the integer which occurs only once in the input array.
lonelyinteger has the following parameter(s):
a: an array of integers

Input Format
The first line contains a single integer, n, denoting the number of integers in the array. 
The second line contains n space-separated integers describing the values in a.

Constraints
1 >= n <100
It is guaranteed that n is an odd number and that there is one unique element.
0 >= a[i] <= 100, where 0 <=i< n.

Output Format
Print the unique integer in the array.