Extra Long Factorials
The factorial of the integer n, written n!, is defined as: n* (n-1) * (n-2)*.....*3*2*1
Function Description
Complete the extraLongFactorials function in the editor below. It should print the result and return.
extraLongFactorials has the following parameter(s):
n: an integer
Note: Factorials of  n>20 can't be stored even in a  64-bit long long variable. Big integers must be used for such calculations. Languages like Java, Python, Ruby etc. can handle big integers, but we need to write additional code in C/C++ to handle huge values.
We recommend solving this challenge using BigIntegers.

Constraints
1 <= n <=100

Sample Input
25
Sample Output
15511210043330985984000000

Note:
Types that exist in Python currently:
Python currently distinguishes between two kinds of integers (ints): regular or short ints, limited by the size of a C long (typically 32 or 64 bits), and long ints, which are limited only by available memory. When operations on short ints yield results that don't fit in a C long, they raise an error.

Why implement both short and long types as int in python?
A new type, integer, is introduced that is an abstract base type of which both the int and long implementation types are subclassed. 
This is useful so that programs can check integer-ness with a single test:
if isinstance(i, integer): ...